feat: 完善文章的增删改查功能

+++ src/main/java/com/isyxf/blog/service/ArticleService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface ArticleService {
     void add(Article article);
-    void remove(Integer id);
+    void remove(long id);
     void edit(Article article);
     Article getById(Integer id);
     List<Article> getAll();
